+++ tools/tools.hpp
@@ -10,7 +10,7 @@
 #include <vector>
 #include "config.hpp"
 
-REAL StrToReal(std::string str);
+double StrToReal(std::string str);
 int StrToInt(std::string str);
 // GSL error handler
 void ErrHandler(const char * reason,
@@ -18,15 +18,15 @@ void ErrHandler(const char * reason,
                         int line,
                         int gsl_errno);
 
-REAL Alpha_s(REAL Qsqr, REAL scaling=1.0);
-REAL Alpha_s_r(REAL rsqr, REAL scaling=1.0);
-REAL Alphabar_s(REAL Qsqr, REAL scaling=1.0); // \alpha_s N_C / Pi
+double Alpha_s(double Qsqr, double scaling=1.0);
+double Alpha_s_r(double rsqr, double scaling=1.0);
+double Alphabar_s(double Qsqr, double scaling=1.0); // \alpha_s N_C / Pi
 
 double T_A(double b, int A);	// Transverse density profile of the nucleus normalized by A
 								// Unit: [b]=1/GeV!
 void InitializeWSDistribution(int A);
 
-int FindIndex(REAL val, std::vector<REAL> &array);
+int FindIndex(double val, std::vector<double> &array);
 
 // Subtract the samallest element of the array from each element
 void SubtractMinimum(std::vector<double> &array);
